# backend/src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Muscians Management - OpenAPI 3.0",
        "description": "One documentation page for all the APIs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": ""
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.11"
    },
    "servers": [
        {
            "url": "http://localhost:3001"
        }
    ],
...
    "paths": {
...
        "/api/user/id/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get registered User by Id",
                "description": "description",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id of desired User",
                        "required": true,
                        "default": "63cef7e50e4632138e25c814"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Got User successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/user/role/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Users by Role",
                "description": "description",
                "operationId": "getUserByRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role of desired User",
                        "required": true,
                        "default": "freelancer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Got User successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "Authorization": []
        }
    ]
}